Renames storage_artifact to artifact
