+++ opsml/projects/active_run.py
@@ -13,7 +13,7 @@ from opsml.registry.cards.run import RunCard
 from opsml.registry.cards.types import METRICS, PARAMS, CardInfo, CardType
 from opsml.registry.sql.registry import CardRegistries, CardRegistry
 from opsml.registry.sql.semver import VersionType
-from opsml.registry.storage.artifact_storage import save_artifact_to_storage
+from opsml.registry.storage.artifact import save_artifact_to_storage
 from opsml.registry.storage.client import StorageClientType
 from opsml.registry.storage.types import ArtifactStorageSpecs
