+++ opsml/registry/cards/run.py
@@ -14,7 +14,7 @@ from opsml.registry.sql.records import (
     RunRegistryRecord,
 )
 from opsml.registry.storage import client
-from opsml.registry.storage.artifact_storage import load_record_artifact_from_storage
+from opsml.registry.storage.artifact import load_record_artifact_from_storage
 from opsml.registry.storage.types import ArtifactStorageSpecs
 
 logger = ArtifactLogger.get_logger()
