+++ opsml/registry/sql/records.py
@@ -17,7 +17,7 @@ from opsml.registry.cards.types import (
     ModelCardUris,
     RegistryType,
 )
-from opsml.registry.storage.artifact_storage import load_record_artifact_from_storage
+from opsml.registry.storage.artifact import load_record_artifact_from_storage
 from opsml.registry.storage.client import StorageClientType
 from opsml.registry.storage.types import ArtifactStorageSpecs
