+++ opsml/registry/storage/artifact.py
